requirements.txt
FlagEmbedding==1.3.4
fuzzywuzzy==0.18.0
nltk==3.9.1
numpy==2.2.3
openai==1.64.0
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
rank_bm25==0.2.2
Requests==2.32.3
scikit_learn==1.6.1
torch==2.5.1
transformers==4.48.0
